# internal/contract/testdata/openapi.baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Regnovum API",
    "version": "0.1.0",
    "description": "Public HTTP contract of Regnovum. Versioned per the master plan: routes registered by the binary must match this document; the contract test fails on drift.",
    "summary": "Arena API contract"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versioned API surface served by the arena binary."
    }
  ],
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "authRegister",
        "summary": "Register a new account",
        "description": "Registers a new account. Prevents user enumeration by returning a uniform confirmation.",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account registration initiated.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegisterResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Validation" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/v1/auth/verify": {
      "get": {
        "operationId": "authVerify",
        "summary": "Verify account email",
        "description": "Verifies an account email using the single-use cryptographic verification token.",
        "tags": ["auth"],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "description": "Single-use raw verification token.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatusResponse" }
              },
              "text/html": {
                "schema": { "type": "string" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Validation" }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "authLogin",
        "summary": "Authenticate and create session",
        "description": "Authenticates credentials and establishes a new session, setting the arena_session cookie.",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Validation" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "authLogout",
        "summary": "Revoke active session",
        "description": "Revokes the active session and clears the arena_session cookie.",
        "tags": ["auth"],
        "security": [{ "SessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Logged out successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatusResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password-reset/request": {
      "post": {
        "operationId": "authRequestPasswordReset",
        "summary": "Request password recovery",
        "description": "Requests a password reset token. Anti-enumeration: returns 200 regardless of account existence.",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PasswordResetRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email dispatched if account exists.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatusResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Validation" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/v1/auth/password-reset": {
      "get": {
        "operationId": "authViewPasswordReset",
        "summary": "Password reset landing page",
        "description": "Renders the password reset HTML form for email recovery links.",
        "tags": ["auth"],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "description": "Password recovery token.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "HTML form rendered.",
            "content": {
              "text/html": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/password-reset/confirm": {
      "post": {
        "operationId": "authConfirmPasswordReset",
        "summary": "Complete password reset",
        "description": "Consumes the password recovery token, updates the credential, and revokes all active sessions.",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PasswordResetConfirm" }
            },
            "application/x-www-form-urlencoded": {
              "schema": { "$ref": "#/components/schemas/PasswordResetConfirm" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset complete.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatusResponse" }
              },
              "text/html": {
                "schema": { "type": "string" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Validation" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
    "/api/v1/me/mfa/enrollment": {
      "post": {
        "operationId": "beginMFAEnrollment",
        "summary": "Begin a second factor enrollment",
        "description": "Generates a pending TOTP secret for the authenticated account and returns it with the otpauth URI exactly once. The secret is stored sealed (AES-256-GCM, bound to the account identifier) and never returned again. The enrollment stays pending until a code confirms it, and an account with a confirmed enrollment is refused, so this route can never replace a working second factor (P16-T05).",
        "tags": ["auth"],
        "security": [{ "SessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Pending enrollment with its secret and otpauth URI.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MFAEnrollmentResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/me/mfa/enrollment/confirm": {
      "post": {
        "operationId": "confirmMFAEnrollment",
        "summary": "Confirm a second factor enrollment",
        "description": "Verifies a code from the pending enrollment, records the confirmation and returns the one-time recovery codes in their only clear-text appearance: they are stored hashed and can be spent exactly once. The code that confirms the enrollment is spent, so it cannot be replayed at step-up (P16-T05).",
        "tags": ["auth"],
        "security": [{ "SessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MFAConfirmRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Enrollment confirmed with its one-time recovery codes.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MFAConfirmResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Validation" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/v1/me/mfa/step-up": {
      "post": {
        "operationId": "stepUpMFA",
        "summary": "Elevate the session with a second factor code",
        "description": "Verifies a TOTP code and marks the calling session as second-factor elevated, which is what the administrative surface requires. The accepted time step is spent, so a replay of the same code is refused even inside the tolerated clock window (P16-T05).",
        "tags": ["auth"],
        "security": [{ "SessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MFAVerifyRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session elevated.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MFAElevationResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Validation" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/v1/me/mfa/recovery": {
      "post": {
        "operationId": "recoverMFA",
        "summary": "Spend a one-time recovery code",
        "description": "Spends one backup code and elevates the calling session: the path of an operator whose authenticator is gone. The code is consumed before anything else and only a code that was never spent can be spent, and the recovery is recorded in the audit trail before the session is elevated, so a recovery that cannot be attributed cannot succeed (P16-T05).",
        "tags": ["auth"],
        "security": [{ "SessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MFAVerifyRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session elevated by a recovery code.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MFAElevationResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Validation" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/v1/me/sessions": {
      "get": {
        "operationId": "listSessions",
        "summary": "List the caller's own sessions",
        "description": "Returns the sessions of the authenticated account that are still usable under the session policy (idle and absolute expiry), most recently seen first, with the calling session marked. The list carries only what each session recorded — creation, last activity, deadline, address and user agent — and never a token, a token hash or a derived device fingerprint (P16-T06). Responses are private, no-store (THR-CACHE-01).",
        "tags": ["auth"],
        "security": [{ "SessionCookie": [] }],
        "responses": {
          "200": {
            "description": "The account's usable sessions.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionListResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/me/sessions/rotation": {
      "post": {
        "operationId": "rotateSession",
        "summary": "Rotate the calling session",
        "description": "Ends the session making the request and issues a fresh opaque token in the same response, so a copy of the previous token taken before the rotation is worthless after it. The new token is generated server-side from the platform's entropy port: no value supplied by the client can become a session token, which is what closes session fixation. Responses are private, no-store (THR-CACHE-01).",
        "tags": ["auth"],
        "security": [{ "SessionCookie": [] }],
        "responses": {
          "200": {
            "description": "The session was rotated; the response carries the new cookie.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionRotationResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 8 }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "type": "string", "enum": ["pending_verification"] }
        }
      },
...
      "PasswordResetRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": { "type": "string", "format": "email" }
        }
      },
...
      "StatusResponse": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "type": "string" }
        }
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "arena_session",
        "description": "Opaque session token cookie (Secure, HttpOnly, SameSite). Sessions arrive with the identity phase; the scheme is declared now to anchor the contract."
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Per-request CSRF token required by state-changing requests once sessions exist."
      }
    },
...
    "responses": {
      "Validation": {
        "description": "Request validation failed.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
...
      "RateLimited": {
        "description": "Too many requests.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
...
    },
...
  },
...
  "security": [],
...
}